Revert "Generic api code added."
This reverts commit 1b3698153913ba44857946bdac37c3bd3fe8b30f.

+++ Youtube MVVM Products/Screens/Product/Model/Product.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Product: Decodable {
     let id: Int
     let title: String
-    let price: Double
+    let price: Float
     let description: String
     let category: String
     let image: String
@@ -18,6 +18,6 @@ struct Product: Decodable {
 }
 
 struct Rate: Decodable {
-    let rate: Double
+    let rate: Float
     let count: Int
 }
